SwiggyRound1/src/test/java/BattleTest.java
package test.java;
import main.java.Battle;
import main.java.Player;

public class BattleTest {

    private Player player1;
    private Player player2;

    @BeforeEach
    public void setUp() {
        // Initialize players with example values
        player1 = new Player(100, 20, 15); // health, strength, attack
        player2 = new Player(120, 18, 12); // health, strength, attack
    }

    @Test
    public void testBattle_Player1Wins() {
        // Set up conditions where player1 wins
        player1.setHealth(100); // Reset health for player1
        player2.setHealth(50);  // Set lower health for player2

        Battle.battle(player1, player2);

        assertTrue(player2.getHealth() <= 0); // Assert that player2 has lost
    }

    @Test
    public void testBattle_Player2Wins() {
        // Set up conditions where player2 wins
        player1.setHealth(80);  // Set lower health for player1
        player2.setHealth(100); // Reset health for player2

        Battle.battle(player1, player2);

        assertTrue(player1.getHealth() <= 0); // Assert that player1 has lost
    }

    @Test
    public void testBattle_Draw() {
        // Test a scenario where both players have equal health
        player1.setHealth(100);
        player2.setHealth(100);

        Battle.battle(player1, player2);

        assertTrue(player1.getHealth() <= 0 || player2.getHealth() <= 0); // Either player could lose or it could be a draw
    }

    private void assertTrue(boolean b) {
        throw new UnsupportedOperationException("Unimplemented method 'assertTrue'");
    }

    // Add more test cases to cover edge cases and different scenarios
}